dont run the execution tests by default
testing the results of the execution is only really necessary when
you're originally writing the 'expected' code output, in order to
make sure that your expected code produces the same results in v2
as the input code produces in v1. once that test passes, there is
no need to continually run that each time. plus its slow