Commit 2 - Fix - Empty 'Random' removal
Removal of Random can cause script lockups if the command is within IF that has no braces.
There are no errors messages or explanations about the lockup from the AHK interpreter... the code simply does not run as expected.
This "fix" adds a filler line that prevents the lockup - the user can remove the filler line after they correct the code (after conversion).